[xulrunner] Use the system NSS CA trust library instead of the bundled copy
With the addition of p11-kit, we want to unify the CA trust stores
across the platform. p11-kit integrates with NSS by replacing
libnssckbi, the CA trust library, with a binary compatible version that
accesses the p11-kit store.

This is implemented in the system-wide nss and p11-kit packages by replacing a
subpackage that contains this library. For this browser version, so long
as we can rely on the ABI being compatible between browser nss, system nss,
and p11-kit, we can simply symlink to the system library and get the
same results.