[embedlite] Use progressive painting by default JB#29001
Per documentation of the feature from: https://wiki.mozilla.org/B2G/TiledLayers

 "As the tiled layers backend forces us, to some extent, to do our drawing updates
  in batches of tiles, it also gives us the opportunity to allow batches to be
  split up and for drawing to be cancelled. This happens on Firefox for Android,
  and is one of the largest performance benefits of the tiles backend."

Unfortunately the feature was still disabled by default when Gecko 31 was
released. It was switched on for release 32. Since the diff betweeen 31
and 32 is not that huge and the necessary patches have been backported lets
switch it on also for embedlite 31. I believe that the benefits
(especially on tablet) outhweight the risks of doing so.