[embedlite] Expose an API for clearing gecko compositing surface. JB#29995
In new sailfish-browser rendering architecture when all the tabs share
the same compositing surface we need the ability to clear it in cerain
cases. One example is when all the tabs get closed. By default this
would mean that the content of the last rendered active page will ocupy
the compositing surface until a new tab is created and some page is
loaded into it. This is rather bad from user experience POV. The problem
could theoretically be solved by clearing the surface on the browser
side, but since gecko draws to it from a dedicated compositor thread it
would require an additional synchronization mechanism to be introduced.

This patch takes a simplier approach by introducing a dedicated API for
clearing view compositing surface with a given color. This should allow
any application to clear the compositing surface whenever it wants
provided the compositor for the view has been initialized. The function
will execute necessary GL commands on an appropriate gecko thread.