[embedlite] Add support for using external compositor GL context.
This implements generic support for using compositor GL context provided
by the application. The aim of the patch is to generalize original Jolla
external window handling patch into something that can be included in
upstream embedlite. Compared to the original implementation this one
does not modify any code in the gecko engine itself. It's all contained
within embedlite layer.

The whole functionality is optional and can be controller by
embedlite.compositor.external_gl_context preference. When set to false
(default) Gecko composior will create GL context on it's own as it
normally does. Setting the pref to true will make embedlite request GL
context from the application through
EmbedLiteViewListener::RequestCurrentGLContext. The implementation of
the funtion is supposed to create the GL context and surface the
compositor is supposed to use and make it current via eglMakeCurrent
call.