[xulrunner] Provide better solution for image decoding problem. Fixes JB#25287.
I believe the new patch actually fixes the root cause behind the bug.
It's not a workaround. Locking images in nsDocument should not automatically
decode all of them in case images.mem.decodeondraw pref is enabled. The
image lock should be established but the actual decoding should happen
during image draw operation.

With this patch in place both the original problem JB#25287 and the
regression it caused (JB#30305) should be fixed.