[embedlite] Add support for setting view margins.
The idea is that the view does not necessarily have to cover the whole
window. An example of such use case is when the browser wants to show
some sort of overlay chrome on top of the view. In such case we don't
want to resize the compositing surface, but we'd like the engine to
avoid painting any content on the area covered by chrome.

Signed-off-by: Piotr Tworek <piotr.tworek@jollamobile.com>