[embedlite] Cleanup the code, make sure Qt/Quick based embed widget works.
This part is accompanied by a similar set of changes in qtmozembed
layer. With both parts in place QOpenGLWebPage and QuickMozView should
work almost the same way as before the refactoring. The main difference
is the top level embedding layer can control compositor handling via the
EmbedLiteWindow object. The Qt4 embedding widgets are no longer
supported.