[embedlite] Make EmbedLiteWindowListener::CompositingFinished more reliable.
The notification was dispatched from custom compositor draw scheduling
code which was only used when EmbedLiteWindowListener::Invalidate
returned false. This means the event was not dispatched for qtmozembed's
QOpenGLWebPage which does not use offscreen rendering path. Instead of
using the custom EmbedLiteCompositorParent::RenderGL function to
dispatch finish events do it from nsIWidget::PostRender.