Add API for setting APZC's DPI
While it is possible to set DPI in EmbedlitePuppetWidget class using
the gfxPlatform::GetDPI() call this aproach doesn't seem to be plausible,
because theoretically the device may have more than one screen with different
DPI. Thus gfxPlatform::GetDPI() is not used neither in B2G nor in Metro.
Instead these two embeddings set DPI at the moment of view creation.
And their PuppetWidget works as a proxy for the real window which lives
in the UI thread. Particularly PuppetWidget forwards GetDPI() requests
from TabChild to TabParent synchrnously. And TabParent maintains its
own copy of DPI value set during window creation.

So, this commit also overrides GetDPI() for EmbedlitePuppetWidget.
First call to this methods gets forwarded synchronously to the UI thread
with the help of EmbedLiteViewThreadChild. The result is cached. After
that only cached value is returned.