[embedlite] Refactor content size and orientation change handling. Contributes to JB#32226.
Many of the ideas implemented by this patch have been taken from my
previous work on the same feature in gecko v38. Basically the rotation
handling is moved entirely to the gecko widget layer. This closely
reassembles what B2G/Gonk is doing. V31 is a bit different in some
regards from v38, but it's close enough.

The main differences from the current v38 work is in unification of size
and rotation change handling. The main problem with rotation and size
changes is in the fact that TabChildBase::HandlePossileViewportChange is
an expensive function to call. We need to call it always when the width
of the viewport might have changed. This can obsiously happen in case
the window is resized (desktop mostly) or rotated. To minimize the
amount of necessary TabChildBase::HandlePossibleViewportChange calls
size and rotation changes were merged into one function. Additionally,
each view dimension change has to be acked by the view child. This
allows the parent to drop pending size requests which have been
invalidated. In previous architecture if the embedder changed view size
quickly 5 times, all the requests were queued and processed (triggering
expensive viewport update 5 times). In the new arch in most cases onl
first and last requests will be processed.

To allow application to show some sort of placeholder when potential
expensive rotation/viewport width change is processed some additional
view notifications have been introduced.