[embedlite] Guard PreRender with visibility and active state. Fixes JB#32879
In case pupper widget is invisible or inactive there is
no need to check rendering status from the owner window. If
the pupper widget is invisible or inactive, LayerManagerComposite
should not render.

When a embedlite view is created set the widget of the window
as active. Similarly active state of the view's widget is updated
when active state changes even thou that doesn't do anything meaningful
at the moment.

Reason for adding active for puppet widget is that visibility
of the puppet widget is overloaded. Furthermore, this makes
controlling of rendering easier.