[embedlite] Fix offscreen rendering pipeline creation. Contributes to JB#35100
When offscreen rendering is enabled the surface size needs to
be known when GLContextProvider::CreateOffscreen is called. If size
is empty, context creation fails and we would create another window instance.

This commit allows us to pass initial window for the EmbedLiteWindow
that will passed to the created Compositor instance from
EmbedLitePuppetWidget::CreateCompositor. This change does not
affect on the external window usage as "embedlite.compositor.external_gl_context"
preference is preferred over MOZ_LAYERS_PREFER_OFFSCREEN environment
variable (see EmbedLitePuppetWidget::GetNativeData and CompositorOGL::CreateContext)

In addition this commit removes mLastViewSize from EmbedLiteCompositorParent
as that will be always the same as mEGLSurfaceSize of the CompositorParent.

Furthermore, setting surface size to zero is know guarded.