[xulrunner] Fix embeded offscreen surfaces not being painted. Contributes to JB#48494
Instead of trying to intercept composite task to redraw to an embedded
surface which stopped working when a scheduling was moved to a distinct
object override the actual composite call. Then only swap the front and
back buffers if a composite task succeeds and do this before notifying
the front end that compositing was done. And finally add a version of
GetPlatformImage which takes a callback to invoke to pass the image so
a lock can be held preventing buffers being swapped while an image
pointer belonging to one of those buffers is being referenced.