[media] Enable GMP extensions. Fixes JB#50322
Modify earlier patch to disable only the heavier parts of the extensions
system that are disabled in Android. This will allow Gecko Media Plugins
to be loaded.

Prioritize GMP decoders in PDMFactory, and implement check for video decoding support in GMPDecoderModule.

Initialize the XUL directory with GRE_HOME, rather than the libxul.so path.

Automatically load gmp-droid at startup if installed