[sailfishos][compositor] Avoid rogue origin points when clipping rects. Contributes to JB#50112
In Matrix4x4Typed::TransformAndClipRect it was possible for the returned
aVerts array to include entries from the initialised but unset
points[][] buffer. As a result, points set to (0, 0) could creep in,
resulting in the distance-to-edge anti-aliasing incorrectly rendering
triangle artefacts.

This change prevents the overrun of the unset points.