[sailfishos][compositor] Allow render shaders to be loaded from file
This change is turned off by default. Set
COMPOSITOR_LOAD_SHADERS_FROM_FILE to enable it (see below).

Ordinarily shaders are generated in-code. This can be cumbersome since a
rebuilt is needed just to edit a shader.

This change will search for shaders in the ~/shaders/ folder, named
using the hex representation of the feature bitfield (e.g. 2080.fs and
2080.vs for the respective fragment and vertex shaders with
ENABLE_TEXTURE_RB_SWAP=0x80 and ENABLE_DEAA=0x2000 enabled; see
the ShaderFeatures enum in OGLShaderProgram.h for the flag values).

Define COMPOSITOR_LOAD_SHADERS_FROM_FILE to enable the code,
for example by adding the following to the %build section of
xulrunner-qt5.spec to enable it.

echo "export CXXFLAGS=\"\$CXXFLAGS \
  -DCOMPOSITOR_LOAD_SHADERS_FROM_FILE \"" >> "$MOZCONFIG"