[sailfishos][media] Fix plugin-container loading from privileged browser. Fixes JB#50399
Gecko tries to add the gre lib dir to LD_LIBRARY_PATH when loading plugin-container, but as sailfish-browser has privileged EGID, glibc removes it for security reasons. Patchelf can insert this into the ELF RPATH instead. libxul also links to liblgpllibs.so in the same directory, so this also needs patching.

This should be reverted if/when the browser no longer needs to run with the privileged egid.