[sailfishos][gecko] Enable startupCache. Contributes to JB#50420
Enables the startupCache for armv7hl builds. Adds
QT_QPA_PLATFORM=minimal to the build environment to allow it to build.

When enabled on i486 the build stalls attempting to run xpcshell to
package the cache contents, so it remains disabled on this architecture.

This also adds functionality to the StartupCache initialisation code to
delete the startupCache if it's older than the timestamp on
/var/lib/_MOZEMBED_CACHE_CLEAN_ .

This _MOZEMBED_CACHE_CLEAN_ file should be touched whenever the packages
are updated to ensure the process works correctly.