Add nsString.h header to GeckoLoader
GeckoLoader was leeching its nsCString implementation from other files
in the unified build process. This adds in the appropriate header so
that it can be built independently.

Identified because adding a new file to the unified build process pushed
GeckoLoader to the top of the next batch, resulting in a build error.