[sailfishos][gecko] Include XUL videocontrols reflow code on Sailfish OS
The desktop browser videocontrol UI is implemented in HTML, whereas on
Sailfish OS (and Android) XUL content is still needed. This change
ensure the XUL code used on Android is also applied for Sailfish OS.

The XUL code is also removed for Android in ESR61. See bug 1510384 and
the linked patch.

https://bugzilla.mozilla.org/show_bug.cgi?id=1510384

https://hg.mozilla.org/mozreview/gecko/rev/9e6a5f9fbde7bd3cf590bc3f8e0736fce170361d