[sailfishos][gecko] Include XUL videocontrols reflow code on Sailfish OS. Contributes to JB#51471
The desktop browser videocontrol UI is implemented in HTML, whereas on
Sailfish OS (and Android) XUL content is still needed. This change
ensure the XUL code used on Android is also applied for Sailfish OS.

The practical consequence of this patch is that the appearance of the
audio playback controls is improved (more neatly integrated into the
page) and the video controls appear in the correct position (along the
bottom of the video rather than the top).

The XUL code is also removed for Android in ESR61. See bug 1510384 and
the linked patch.

https://bugzilla.mozilla.org/show_bug.cgi?id=1510384

https://hg.mozilla.org/mozreview/gecko/rev/9e6a5f9fbde7bd3cf590bc3f8e0736fce170361d