[sailfishos][gecko] Make <button> hit testing similar to other elements. Contributes to JB#51683
This is a backported version of the fix for mozilla bug 1089326: "make
<button> hit testing similar to other elements which may have some
content, and for click target find the common (interactive) ancestor,
r=masayuki"

See https://hg.mozilla.org/mozilla-central/rev/36b949bef798

Essentially it makes hit targetting of the <button> element the same as
for other elements. Previously only the button element itself could be
the target, even for hits on its children. This change makes it so that
the children can be targets as well (so select, hover, etc. can be
captured by them as well).

This patch should be removed after the upgrade to ESR68.