[sailfishos][embedlite] Cleanup embedding/embedlite/Makefile.in DEFINES. JB#53003
The error is following:
Exception: Variable DEFINES is defined in gecko-dev/mobile/sailfishos/Makefile.in.
It should only be defined in moz.build files.

Once we get build moving forward we need to also cleanup DEFINES from
embedding/embedlite/installer/Makefile.in (see JB#53017) and might be
other issues lurking in the packaging.