[sailfishos][embedlite] Allow current window id to be retrieved. JB#54218 OMP#JOLLA-135
When a new WebView is created EmbedLite increments the view ID. This ID
is needed when initialising the new compositor, but currently there's no
way to find out which ID to use.

This change allows the last created view ID to be retrieved for use
during compositor initialisation.

Cherry-pick of sha1 37b9769b5668652004eeb8bc4780a36b8b3e2147