[sailfishos][gecko] Update CompositorManagerParent patch to use the correct view ID. JB#54218 OMP#JOLLA-135
Updates the CompositorManagerParent patch so that it passes in 0 as the
view ID, rather than 1. When using 0 the latest ID is used, ensuring
that the compositor is associated with the correct EmbedLiteWindow
(which is the latest one to be created).

This prevents a segfault when closing and recreating a WebView.