[sailfishos][embedlite] Pass EmbedLiteWindowListener pointer to PEmbedLiteWindowConstructor. Contributes to JB#54813
nsWindow::GetGLContext is called by SendPEmbedLiteWindowConstructor
asynchronously. If this happens before the EmbedLiteWindow is added to
EmbedLiteApp::mWindows, EmbedLiteApp::GetWindowByID returns nullptr and
the browser draws white window. Similarly, if nsWindow::GetGLContext is
called before EmbedLiteWindow::SetListener, EmbedLiteWindow::GetListener
returns a pointer to the FakeWindowListener and the browser draws white
window. To prevent this, the EmbedLiteWindowListener pointer should be
passed to EmbedLiteWindowBaseChild directly.