[sailfishos][configure] Patch libloading to build on arm.
This patch overrides the default build step using cc::Build to force the
right tool to be used (host-gcc) and allow libloading to be compiled and
linked against.

It also introduces the SB2_TARGET environment variable set to the
architecture so that the rust build can adjust its approach accordingly.

SB2_TARGET takes one of the following values:

1. armv7-unknown-linux-gnueabihf
2. aarch64-unknown-linux-gnu
3. i686-unknown-linux-gnu