[sailfishos][fonts] Load and set FTLibrary for the Factory. Fixes JB#55191 OMP#JOLLA-285
The same old FTLibrary loading issue but another incarnation of it.
The gfxFontGroup::GetDefaultFont of the gfxTextRun failed to
FindOrMakeFont gfxFont and set it as default font. The gfxFontEntry::FindOrMakeFont
itself failed to CreateFontInstance via gfxFontconfigFontEntry::GetFTFace
which fails because FTLibrary was not loaded by the platform. See more details from the
gfx/2d/Factory::NewSharedFTFace.

This commit adds FreeType library loading and setting up the
library for the Factory.