[sailfishos][configure] Expose sb2 target to build process
Custom build targets require special workarounds when cross-compiling.
This change adds the SB2_TARGET environment variable which can be read
by the build.rs code in order to determine whether the sb2 target is
different from the rust build target. If it is, the workaround can be
applied.

Possible values for SB2_TARGET are currently:

1. armv7-unknown-linux-gnueabihf
2. aarch64-unknown-linux-gnu
3. i686-unknown-linux-gnu