[sailfishos][embedlite] Fix delivery of MozAfterPaint events. JB#55286 OMP#JOLLA-310
Revert a patch to remove an ensure eventual after paint event timer as
no event is delivered in embedlite without it.

The short explanation is that the root widget (an embedlite puppet
widget) of the PresContext where the event is scheduled, is not the same
root widget as in the ClientLayerManager (nsWindow) where the paint is
completed, so the nsView installed as the widget listener on the puppet
widget doesn't have the DidCompositeWindow called via
ClientLayerManager::DidComposite. And even with a redirect established
there's a subsequent failure because the DidComposite is a posted event
and the transaction it refers to is gone when it executes.