[sailfishos][gecko] Reapply patch: "Get 12/24h timeformat setting from dconf". JB#55228 OMP#JOLLA-298
This reapplies the ESR60 patch to the ESR78 code. Some of the changes
were in the embedding portion of the project; these have been moved out
of the patch and applied directly.

Like many other Sailfishos DConf settings, selecting between 12h and
24h timer formats involves key containing underscore in name. Such keys
can't be used in GSettings schemas and thus are incompatible with the
already existing lookup logic.

If sailfish setting for timeformat exists, access and use it directly
via dconf API. Otherwise fall back to default behavior involving
GSetting schemas.