[sailfishos][gecko] Reapply patch: "Ensure audio continues when screen is locked". JB#55228 OMP#JOLLA-298
This reapplies the ESR60 patch to the ESR78 code.

[sailfishos][media] Ensure audio continues when screen is locked.
Contributes to JB#51747

This rationalises two previous commits from sailfishos-esr52,
integrating the functionality of the NemoResourceHandler class into
MediaDecoder.

This is needed so that the "media-decoder-info" notifications are sent
out when certain types of media are started/stopped and to signal
whether the media contains audio and/or video.

An additional change ensures media is started/stopped with the
Thaw/Freeze of docuemnts (switching the check from visible status to
hidden status).

The previous relevant commits:-

commit 9e2dc6c89b83a623ad27ac14edc501585ef8ecd9
Author: Oleg <romaxa@gmail.com>
Date:   Thu Oct 30 23:54:05 2014 -0700

    [sailfishos][media] Add NemoResourceHandler

    Original SHA1: b318b3c9f723544c32895a8a4c49d7f44b6752be

    Signed-off-by: Raine Makelainen <raine.makelainen@jolla.com>

commit 50fc4fac5ed1433acb0dff6e23dc3055f76c3226
Author: Andrew den Exter <andrew.den.exter@qinetic.com.au>
Date:   Thu Nov 29 17:20:06 2018 +1000

    [sailfishos][media] Update media resource state when the playing
state changes for any reason. Fixes JB#42900