[sailfishos][gecko] Reapply patch: "Hardcode loopback address for profile lock filename". JB#55228 OMP#JOLLA-298
This reapplies the ESR60 patch to the ESR78 code.

Hardcode loopback address for profile lock filename. JB#45654

When engine started without network PR_GetHostByName takes 20seconds
when connman tries to resolve host name. As this is only used as part of
the profile lock filename it can as well be like "127.0.0.1:+<pid>".