[sailfishos][gecko] Reapply patch: "Fix flipped FBO textures when rendering to an offscreen target". JB#55228 OMP#JOLLA-298
This reapplies the ESR60 patch to the ESR78 code.

Fix flipped FBO textures when rendering to an offscreen target.

Only take whether the GL context is for an offscreen surface into
account when rendering to the window surface. Internal FBO surfaces
should behave the same irrespective of the final render target for the
web page.