[sailfishos][embedlite] Use StaticXREAppData to read application.ini. Fixes JB#55883 OMP#JOLLA-436
Quote from XREAppData.h:
"A static version of the XRE app data is compiled into the application
so that it is not necessary to read application.ini at startup.

This structure is initialized into and matches nsXREAppData"

If there are issues in future those should be handled by build config and
fixing application.ini generation.

This also marks GetSourceURL and GetUpdateURL as ok and not assigning
anything there.