[sailfishos][gecko] Fix media elements being suspended on the active tab. JB#56081 OMP#JOLLA-488
HTMLMediaElement::IsHidden() was being used instead of IsActive to determine
if the element was not in the active tab and should therefore be
suspended. But IsHidden() also returned true if IsInComposedDoc() was
false which meant all elements not in composed document were forever
suspended. Using just the hidden state of the owner document instead
removed that condition.