[sailfishos][gecko] Fix download starting when multiple tabs are closed. JB#56047 OMP#JOLLA-476
When the active tab is closed another one is made active, and content is
downloaded. If that new tab is also then closed there is a window where
the download may complete after the view for the tab is destroyed but
before all references to the document are. A new content viewer cannot
be constructed for the downloaded document because the parent widget has
been invalidated, this is misinterpreted as the document being
unsupported for display and a fallback save to disk is triggered
instead.

Forcing destruction of the nsDocShell by destroying the nsWebBrowser
before destroying the parent widget will put it into a state where it
will cancel any further downloads it is asked to handle.