[sailfishos][embedlite] Make EmbedFrame an EventListener. JB#56094 OMP#JOLLA-492
A message handler in RemotePageHandler was failing because it attempted
to install an event handler on the target of the received message but
addEventListener wasn't a didn't exist on EmbedFrame. The equivalent
object in Firefox appears to MozBrowser which is a javascript defined
object so there isn't a direct translation, but similarly structured
interfaces appear to be defined in webidl and extend EventTarget.