[sailfishos][embedlite] Supply the correct target when delivering messages. JB#56094 OMP#JOLLA-492
A message handler in RemotePageHandler was failing because it attempted
to install an event handler on the target of the received message but
addEventListener wasn't a didn't exist on EmbedFrame. Other code paths
were already setting the message manager as the target so this brings
everything into the one model.