[sailfishos][embedlite][egl] Fix egl patch to prevent crashing. Fixes JB#56606 OMP#JOLLA-565
A static function in the code had return value bool, but no return
statement. The compiler then optimised this undefined behaviour into a
crash (calling an otherwise guarded dlopen with an invalid parameter).

This change removes the crash by setting the return as void. Now the
dlopen is only called when it's supposed to be.

Thanks to Tomi Leppanen and especially Simo Piiroinen for noticing the
incorrect code.

This change also fixes up an incorrectly named variable.