[sailfishos][embedlite] Pass double taps to embedhelper.js. JB#56765 OMP#JOLLA-607
Ensure that double taps are passed on to the embedhelper.js, so they can
be passed on to any javascript handlers.

If there are no double click javascript handlers, try to zoom in to/out
from the element instead.

Set the decision about whether to zoom on double tap to follow the
decision made by nsLayoutUtils::AllowZoomingForDocument().