[sailfishos][embedlite] Add a content controller stack. Fixes JB#57690
When multiple WebView componennts are used in an app, since there is
only one compositor, only one can be in direct use at a time. This will
be the last that was set as active.

When the currently activated component is deactivated, the WebView that
was active prior to that should be reinstanted. In particular, the
view's GeckoContentController needs to be reinstated.

This change adds a stack for managing the content controllers, so that
previously active controllers can be reactivated.

Co-authored-by: Raine Mäkeläinen <raine.makelainen@jolla.com>