[sailfishos][gecko] Add support for prefers-color-scheme. JB#58394
Adds support to qt/nsLookAndFeel for the prefers-color-scheme media tag.
The renderer already supports it by setting the ui.systemUsesDarkTheme
preference to 0 or 1.

This patch allows, in addition, for the colour scheme to be controlled
by sending an "ambience-theme-changed" notification with a data payload
of either "light" or "dark". If the ui.systemUsesDarkTheme preference is
unset or set to 2, the notification value will be used to control the
theme instead.

This functionality is used to update the colour scheme based on the
ambience of the device.