Add freetype2 include directory
The include used to be added from the following file, but this no longer
seems to be included:

gecko-dev/media/webrtc/trunk/peerconnection_client.target.mk

To work around this we add it in to the CFLAGS export. There's probably
a much better place to put it, but this seems to do the trick for now.