Reduce optimisation to O1
Without this building swgl fails with the following error:

during RTL pass: expand
src/glsl.h: In function ‘glsl::vec2_scalar glsl::sign(glsl::vec2_scalar)’:
src/glsl.h:662:39: internal compiler error: Segmentation fault
 float sign(float a) { return copysignf(1.0f, a); }
                               ~~~~~~~~~^~~~~~~~~

It's not clear what the exact issue is here, but dropping the
optimisation allows the build to pass. If this is the only way, it would
be good to improve this by having the change only apply to the one
component.