Hide SetIsActive() aActiveId parameter
The nsFocusManager::GenerateFocusActionId() method needed to set the
aActionId parameter isn't part of the libxul public API, so we need to
apply it inside the library.

This change therefore hides the parameter from the public API and sets
this value internally instead.

This change can be combined with commit
21d6b7886c95440bf5604388b96c7e60f08a9216 "Add aActionId parameter to
SetIsActive".