Build with only a single process
Local builds with only a single process, as opposed to multiple
simultaneous processes, have exhibited no tendendancy to stall. This
change therefore switches to using only one build process in the hope of
increasing build robustness, not just locally, but also on OBS.

This is only an interim fix until the underlying cause has been
resolved.